Translate less data in MIR
